Add state and error to profile API
We want to make it easy for customers and oncalls to identify a detector’s state and error if any. This PR adds such information to our new profile API.

We expect three kinds of states:
-Disabled: if get ad job api says the job is disabled;
-Init: if anomaly score after the last update time of the detector is larger than 0
-Running: if neither of the above applies and no exceptions.

Error is populated if error of the latest anomaly result is not empty.

Testing done:
-manual testing during a detector’s life cycle: not created, created but not started, started, during initialization, after initialization, stopped, restarted
-added unit tests to cover above scenario