Fix retrying saving anomaly results
Previously, AD throws an exception when retrying saving an anomaly result, indicating autoGeneratedTimestamp should not be set. We can see the exception because we reuse the same IndexRequest, where autoGeneratedTimestamp is already set. This PR creates a new IndexRequest by copying the original request without autoGeneratedTimestamp.

This PR also adds detector id to AnomalyDetectionException’s toString method so that we can see the id in the log. This PR also adds more logging to SearchFeatureDao.