Choose the number of primary shards while creating indices
AD is bottlenecked by the number of primary shards of job, result, and checkpoint index in HC.  The number of primary shards in the job index determines how many nodes can run as AD's coordinating nodes.  The number of primary shards in the result and checkpoint index determines the extent of index pressure given the same indexing workload. 

Previously, we used the default setting: in ODFE, the number is 1; in AES, the number is 5. This PR uses the number of hot nodes as the number of primary shards for the checkpoint, result, and job index . The upper limit is 10.

Testing done:
1. added unit tests.
2. end-to-end testing