Add caching
In a single-entity detector, the amount of memory used is proportional to the number of features and is fixed when the detector is defined. However, with the introduction of multi-entity detectors, the number of entities is not fixed, and active entities are changing. Thus, the size of the required memory may constantly change in the lifetime of a detector. Caching can accommodate such requirements without the need to pre-allocate memory for a detector in a fixed number. If the memory is enough, we create models for all entities and monitor anomalies. Otherwise, we cache hot entities in memory and make a best-effort attempt to admit other entities.

This PR adds caching for multi-entity detectors.  We have a three-level cache that stores entity states in each node. The first two levels are for active entities. The third level is for inactive entities. Each detector has its dedicated cache that stores ten (dynamically adjustable by humans) entities’ states per node. A detector’s hottest entities load their states in the dedicated cache. If more than 10 entities use the dedicated cache, the secondary cache can use the rest of the free memory available to AD. The secondary cache is a shared memory among all detectors for the long tail. The shared cache size is 10% heap minus all of the dedicated cache consumed by single-entity and multi-entity detectors. The shared cache’s size shrinks as the dedicated cache is filled up or more detectors are started.

Testing done:
1. added unit tests.
2. end-to-end testing