Add memory tracker
Previously, when creating a model, we evaluate all existing models and compare the total with the 10% heap memory limit.  If yes, we proceed to create the model.  Otherwise, we throw exceptions.  This does not work for multi-entity detectors.  First, there can be a lot of models in cache.  Reevaluating them every time we want to add a model is not efficient.  Second, we have two sources of memory usage now: single-entity and multi-entity detectors.  We need a central place to track memory usage across the board as we add more and more kinds of detectors.  This PR achieves the purpose.

This PR also updates RCF model size estimation.  Previously, we underestimated the size.
This PR also adds threshold model size estimation.  Previously, we didn't consider it.

This PR also adds a customized hashmap that can automatically consume and realese memory.  This enables minimum change to our single-entity code as we just have to replace the map implementation.

Testing done:
1. will add unit tests.
2. end-to-end testing pass.