Rename TransportState to NodeState
Previously, we have TransportState and TransportStateManager to track states used by transport layers.  Now the state is not only used by the transport layer. Methods like getDetector are used by ModelManager etc.  This PR renames the class to reflect the fact.

This PR also modifies how we track whether a cold start is running or not.  Previously, the caller had to manually set it on and off.  And we have a code everywhere.  Now, we return a Releasable object that can be called automatically.  The current way is more concise and easier to avoid bugs.

This PR also adds states to track last index throttling time as we face index rejections issues.

Testing done:
1. will add unit tests.
2. end-to-end testing pass.