Add "main" function for model training and inference
This PR adds the "main" function for model training and inference.  If the model is in cache, the function asks the model directly for anomaly grade.  If the model is not in cache, the function triggers a cold start.

This PR also splits the model partition related functions to a separate class to break the circular dependency between ModelManager and NodeStateManager.  This breakup also has better encapsulation as we have one class for one thing.

Testing done:
1.  Will add unit tests.
2. Manual tests pass.