Add rcf score in ThresholdingResult
This PR adds an rcf score in ThresholdingResult so that we can use it when saving anomaly results.  Previously we don't need it since AnomalyResultTransportAction runs rcf first
and then threshold model. But for multi-entity result transport action, it sends the da
ta and gets back the ThresholdingResult, nothing in between.

Testing done:
1. Its callers have covered its usage in their unit tests.
2. Manual testing passes.